+++ tests/Fixtures/Mesh/Forms/Input.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Mesh\Forms;
+
+use EthanBarlo\Mesh\MeshComponent;
+
+class Input extends MeshComponent
+{
+}